requirements.txt
agno
openai
python-dotenv
duckduckgo-search
pypdf
lancedb
streamlit
pandas
tantivy
yfinance
groq
ddgs